requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
absl-py=1.0.0=pypi_0
ca-certificates=2021.10.26=h06a4308_2
cachetools=4.2.4=pypi_0
certifi=2021.10.8=py38h06a4308_0
charset-normalizer=2.0.9=pypi_0
cycler=0.11.0=pypi_0
fonttools=4.28.5=pypi_0
google-auth=2.3.3=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
grpcio=1.43.0=pypi_0
idna=3.3=pypi_0
importlib-metadata=4.10.0=pypi_0
kiwisolver=1.3.2=pypi_0
ld_impl_linux-64=2.35.1=h7274673_9
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgomp=9.3.0=h5101ec6_17
libstdcxx-ng=9.3.0=hd4cf53a_17
markdown=3.3.6=pypi_0
matplotlib=3.5.1=pypi_0
ncurses=6.3=h7f8727e_2
numpy=1.21.5=pypi_0
oauthlib=3.1.1=pypi_0
openssl=1.1.1l=h7f8727e_0
packaging=21.3=pypi_0
pandas=1.3.5=pypi_0
pillow=8.4.0=pypi_0
pip=21.2.4=py38h06a4308_0
protobuf=3.19.1=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pyparsing=3.0.6=pypi_0
python=3.8.12=h12debd9_0
python-dateutil=2.8.2=pypi_0
pytz=2021.3=pypi_0
pyyaml=6.0=pypi_0
readline=8.1=h27cfd23_0
requests=2.26.0=pypi_0
requests-oauthlib=1.3.0=pypi_0
rsa=4.8=pypi_0
scipy=1.7.3=pypi_0
seaborn=0.11.2=pypi_0
setuptools=58.0.4=py38h06a4308_0
six=1.16.0=pypi_0
sqlite=3.37.0=hc218d9a_0
tensorboard=2.7.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.0=pypi_0
tk=8.6.11=h1ccaba5_0
torch=1.8.1+cu111=pypi_0
torchaudio=0.8.1=pypi_0
torchvision=0.9.1+cu111=pypi_0
typing-extensions=4.0.1=pypi_0
urllib3=1.26.7=pypi_0
werkzeug=2.0.2=pypi_0
wheel=0.37.0=pyhd3eb1b0_1
xz=5.2.5=h7b6447c_0
zipp=3.6.0=pypi_0
zlib=1.2.11=h7f8727e_4
